requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyascii==0.3.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
dataclasses-json==0.5.14
defusedxml==0.7.1
Django==4.2.3
django-cogwheels==0.3
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
emoji==2.8.0
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filetype==1.2.0
frozenlist==1.4.0
greenlet==2.0.2
html5lib==1.1
idna==3.4
joblib==1.3.2
l18n==2021.3
langchain==0.0.281
langsmith==0.0.33
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pydantic==2.3.0
pydantic-core==2.6.3
pypdf==3.15.5
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
tabulate==0.9.0
telepath==0.3.1
tenacity==8.2.3
tqdm==4.66.1
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
unstructured==0.10.12
urllib3==2.0.4
wagtail==5.0.2
wagtailmenus==3.1.8
webencodings==0.5.1
whitenoise==6.5.0
Willow==1.5.1
yarl==1.9.2
